+++ ncgen/util.h
@@ -57,7 +57,7 @@ extern int findlastunlimited(Dimset* dimset);
 extern int countunlimited(Dimset* dimset);
 
 extern unsigned char* makebytestring(char* s, size_t* lenp);
-extern int getpadding(int offset, int alignment);
+extern size_t getpadding(size_t offset, size_t alignment);
 
 extern void check_err(const int stat, const int line, const char* file, const char* func);
 extern void check_err2(const int stat, const int cdlline, const int line, const char* file, const char* func);
